Extract user email services into a separate file
