+++ apps/users/services/__init__.py
@@ -1,1 +1,2 @@
 from .base import *
+from .email import *
