+++ apps/users/services/base.py
@@ -3,21 +3,12 @@ from typing import Optional
 
 from allauth.account.models import EmailAddress
 from allauth.socialaccount.models import SocialAccount
-from django.core.exceptions import PermissionDenied, ValidationError
-from django.core.mail import EmailMultiAlternatives
-from django.core.validators import validate_email
+from django.core.exceptions import ValidationError
 from django.db import connection, transaction
 from django.db.models import Count
 from django.db.models.query import QuerySet
-from django.http import HttpRequest
-from django.template.loader import render_to_string
-from django.urls import reverse
-from django.utils.encoding import force_bytes
-from django.utils.http import urlsafe_base64_encode
 
 from users.models import Profile, User
-
-from .tokens import activation_token_generator, email_change_token_generator
 
 
 logger = logging.getLogger("default_logger")
@@ -78,43 +69,6 @@ def toggle_user_supscription(user: User, author: User) -> None:
             author.profile.subscribers.add(user)
         else:
             author.profile.subscribers.remove(user)
-
-
-def send_account_activation_email(request: HttpRequest, user: User):
-    encoded_user_id = urlsafe_base64_encode(force_bytes(user.pk))
-    token = activation_token_generator.make_token(user)
-    url = request.build_absolute_uri(
-        reverse("account-activate", args=[encoded_user_id, token])
-    )
-
-    context = {"username": request.user.get_username(), "url": url}
-    html_content = render_to_string("users/emails/activation_email.html", context)
-    text_content = render_to_string("users/emails/activation_email.txt", context)
-
-    email = EmailMultiAlternatives(
-        subject="User account activation", body=text_content, to=[user.email]
-    )
-    email.attach_alternative(html_content, "text/html")
-    email.send()
-
-
-def send_email_change_link(request: HttpRequest, new_email: str) -> None:
-    if not request.user.is_authenticated:
-        raise PermissionDenied("Login required to change email address.")
-
-    validate_email(new_email)
-
-    token = email_change_token_generator.make_token(request.user)
-    url = request.build_absolute_uri(reverse("email-change-confirm", args=[token]))
-    context = {"username": request.user.get_username(), "url": url}
-    html_content = render_to_string("users/emails/email_change.html", context)
-    text_content = render_to_string("users/emails/email_change.txt", context)
-
-    email = EmailMultiAlternatives(
-        subject="Confirm email change", body=text_content, to=[new_email]
-    )
-    email.attach_alternative(html_content, "text/html")
-    email.send()
 
 
 def enforce_unique_email_type_per_user(instance: EmailAddress) -> None:
